Finished to exercise Part4 Bloglist 4.23

+++ part4/bloglist/controllers/blogs.js
@@ -1,12 +1,14 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const { userExtractor } = require('../utils/middleware')
+
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
 
   if (!body.title) {
@@ -19,15 +21,48 @@ blogsRouter.post('/', async (request, response) => {
     })
   }
 
-  const blog = new Blog(body)
+  const user = request.user
+
+  if(!user){
+    return response.status(401).json({
+      error: 'JWT token is missing or wrong'
+    })
+  }
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    user: user._id
+  })
+
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if(!user){
+    return response.status(401).json({
+      error: 'JWT token is missing or wrong'
+    })
+  }
+
+  if(blog.user.toString() === user._id.toString()){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(400).json({
+      error: 'Only the creator of the blog can delete it'
+    })
+  }
+
 })
 
 blogsRouter.put('/:id', (request, response, next) => {
